Updated autocomplete_user_search.js to better handle labels and tooltips.
Added code to make the new search box take on the id of the one it replaces
It now also moves any labels that reference the replaced search field into the
span so that it can reference the new search field instead.

Finally, the comments were updated to be more useful and informative.